Value Classes based Scala-integration
Aim of this is to replace the Scala support built for Vert.x 3 with something less allocation heavy and to rely on less/no reflection calls.
Value classes allow to define additional behavior for classes and can be used instead of going the way of inheritance or wrapping.
This curretn implementation focuses on making the API nullsafe using Optional and using Scala-Futures for all async parts.
It also adds shorthands for DataObjects using Scala objects.